feat: additional label helper specific to go language
Added additional [PR labeler](https://github.com/joelanford/go-apidiff). This uses golang's experimental
package [api-diff](https://pkg.go.dev/golang.org/x/exp/apidiff).

The idea with this implementation is that, even if the PR is missing context in regards to if
it is a feature, bugfix or a breaking change, then we can check the implementation and see if
it introduces anything new that has been exported, or anything exported has been removed, or
maybe it's a change that doesn't effect the user of the project(not functionally, but aesthetically)