feat: allow ignoring of major changes if path matches pattern
Context:

We have had issues in our internal projects where release-drafter-go
was being used to automatically tag go projects, where the
release-drafter tagged the PR as major changes because
it was trying to introduce breaking changes to a path, `*/v1beta1.*`

We have agreed on that specific project, that breaking changes are allowed
for v1beta1 beta however not allowed for v1 APIs.

We needed a mechanism in release-drafter-go that allowed us to ignore
major changes to path matching a pattern, and simply tag it as a minor
bump.