feat: Diff changes since last tag (release)
This change addresses the false positive detection of breaking changes
when PRs break a change that is not yet part of a release. This should
remove some accidental major version bumps and allow use to
revert/change things before release, but after a change is merged to
main. A similar breaking policy has been put in place in
[github.com/services-interfaces].

[github.com/services-interfaces]: https://github.com/coopnorge/services-interfaces/pull/660